word-game-backend/requirements.txt
flask
gunicorn
azure-cosmos
python-dotenv
flask-cors